Refactor ts and lar message converters
